refactor global state singleton location
