+++ src/errorHandling/errorPrinter.java
@@ -1,7 +1,7 @@
 package src.errorHandling;
 
 import src.errorHandling.types.abstract_error;
-import src.parseTree.nodes.global_state_singleton;
+import src.global_state_singleton;
 
 import java.io.File;
 import java.io.IOException;
