+++ src/global_state_singleton.java
@@ -1,4 +1,4 @@
-package src.parseTree.nodes;
+package src;
 
 import java.util.HashMap;
 import java.util.Map;
